Move shared React hooks into hooks folder
