+++ src/web-v2/api/client.ts
@@ -1,5 +1,5 @@
 export {
   trpcReact,
 } from '@/client-shared/api/react';
-export { useControlPlaneTrpcClient } from '@/client-shared/api/useControlPlaneTrpcClient';
+export { useControlPlaneTrpcClient } from '@/client-shared/hooks/useControlPlaneTrpcClient';
 export type * from '@/client-shared/api/types';
